pyproject.toml
[tool.poetry]
name = "gap-analysis"
version = "0.1.0"
description = "Generating kg with llm"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
alembic = "1.12.0"
beautifulsoup4 = "~=4.13.3"
deprecated = "~=1.2.14"
fastapi = "~=0.100.0"
google-genai = "~=1.5.0"
openai = "~=1.63.0"
pydantic = "~=2.7.4"
pygments = "2.16.1"
psycopg2 = "2.9.9"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-magic = "0.4.27"
pyyaml = "6.0.1"
regex = "2023.8.8"
requests = "2.31.0"
sqlalchemy = "~=2.0.14"
sqlmodel = "~=0.0.22"
starlette = "~=0.27.0"
tiktoken = "~=0.7.0"
uvicorn = "~=0.23.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = true
exclude = '''
/(
    \.git
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
